Move attribute value messages to the separate message folder
