+++ lib/s3i_flutter.dart
@@ -7,7 +7,7 @@ export 'src/auth/oauth_proxy_flow.dart';
 export 'src/auth/tokens.dart';
 
 //broker
-export 'src/broker/attribute_value_messages.dart';
+export 'src/broker/messages/attribute_value_messages.dart';
 export 'src/broker/broker_interfaces.dart';
 export 'src/broker/broker_rest_connector.dart'
     if (dart.library.io) 'src/broker/broker_amqp_connector.dart';
